Refactor scanFace to conditionally render different content
