+++ src/components/Manual.tsx
@@ -1,0 +1,3 @@
+export default function Manual() {
+  return <h1>manual</h1>;
+}
