+++ src/components/MatchingPatients.tsx
@@ -1,0 +1,3 @@
+export default function MatchingPatients() {
+  return <h1>matching patients placeholder</h1>;
+}
